New plugin: degoogle
The purpose of this plugin is to extract the target URL of Google result
link. If a user posts a link copied from Google in a channel, you can
run the .degoogle command do "degoogle" the last posted link:

    <raek> ZOMG, typinferens!!! http://www.google.se/url?sa=t&rct=j&q=&esrc=s&source=web&cd=1&cad=rja&uact=8&ved=0CCEQFjAAahUKEwil0eP444rGAhWBjCwKHb4iAJ4&url=http%3A%2F%2Fciteseerx.ist.psu.edu%2Fviewdoc%2Fdownload%3Fdoi%3D10.1.1.65.7733%26rep%3Drep1%26type%3Dpdf&ei=VCR7VaXXGIGZsgG-xYDwCQ&usg=AFQjCNEZd7yLYNia8S_uC0CF-s0v5PHQ7A&sig2=rOx-Me30SoVfg9zhGdCADg&bvm=bv.95515949,d.bGg
    <not-raek> .degoogle
    <pynik> http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.65.7733&rep=rep1&type=pdf

And everyone is happy. Yay. The .degoogle command can also be given a
URL as an argument.